voxel and bev both pass tests

+++ physics_atv_visual_mapping/localmapping/voxel/voxel_localmapper.py
@@ -121,7 +121,7 @@ class VoxelGrid:
         """
         Get indexes for positions given map metadata
         """
-        gidxs = torch.div((pts[:, :3] - self.metadata.origin.view(1,3)), self.metadata.resolution.view(1,3), rounding_mode='floor')
+        gidxs = torch.div((pts[:, :3] - self.metadata.origin.view(1,3)), self.metadata.resolution.view(1,3), rounding_mode='floor').long()
         mask = (gidxs >=  0).all(dim=-1) & (gidxs < self.metadata.N.view(1,3)).all(dim=-1) 
         return gidxs, mask
